pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurum"
version = "0.1.0"
description = "Shared code and tooling for the Aurum market intelligence platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Aurum Team", email = "engineering@aurum.local"}
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "boto3>=1.34",
  "numpy>=1.26",
  "pandas>=2.2",
  "python-dateutil>=2.8",
  "requests>=2.32",
  "psycopg[binary]>=3.2"
]

[project.optional-dependencies]
ingest = [
  "openpyxl>=3.1",
  "pyarrow>=16.0",
  "xlrd>=2.0",
  "kafka-python>=2.0"
]
testing = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "fastavro>=1.9.4"
]
quality = [
  "ruff>=0.5",
  "mypy>=1.10"
]
iceberg = [
  "pyiceberg>=0.16.0",
  "pyarrow>=16.0"
]
api = [
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "trino>=0.328.0",
  "redis>=5.0",
  "prometheus-client>=0.20",
  "httpx>=0.25",
  "python-jose[cryptography]>=3.3"
]

[project.urls]
"Documentation" = "https://github.com/aurum/dev-docs"
"Source" = "https://github.com/aurum/platform"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.coverage.run]
branch = true
source = ["aurum"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false
